Split the server.js file up into their own files
