+++ Panel/bot/discord/commands/server/move.js
@@ -1,0 +1,3 @@
+exports.run = async (client, message, args, cooldown) => {
+    message.channel.send('Command coming soon! almost finished :D')
+}
